# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProviEmplea API",
        "description": "API REST para la plataforma de empleo ProviEmplea de Providencia.\n    Gestiona talentos, empresas y procesos de selección con curriculum ciego.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo local (Docker)"
        }
    ],
    "paths": {
        "/admin/contactos": {
            "get": {
                "tags": [
                    "Administración"
                ],
                "summary": "Listar solicitudes de contacto",
                "description": "Retorna solicitudes con empresa y persona embebidos. Filtrable por estado.",
                "operationId": "listarContactos",
                "parameters": [
                    {
                        "name": "estado",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pendiente",
                                "contactado",
                                "entrevista",
                                "seleccionado",
                                "no-seleccionado",
                                "proceso-cerrado"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de contactos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ContactoSolicitado"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Administración"
                ],
                "summary": "Registrar solicitud de contacto",
                "description": "Una empresa solicita contactar a un talento. No puede existir solicitud activa previa.",
                "operationId": "crearContactoSolicitado",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContactoInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Solicitud creada en estado pendiente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ContactoSolicitado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Ya existe solicitud activa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorGeneral"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "IDs inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidacion"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/admin/contactos/{id}/estado": {
            "patch": {
                "tags": [
                    "Administración"
                ],
                "summary": "Actualizar estado del proceso de selección",
                "description": "Avanza el estado. Las fechas (contacto, entrevista, resultado) se registran automáticamente.",
                "operationId": "actualizarEstadoContacto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EstadoInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estado actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ContactoSolicitado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/admin/estadisticas": {
            "get": {
                "tags": [
                    "Administración"
                ],
                "summary": "Estadísticas generales de la plataforma",
                "description": "Resumen para el dashboard del equipo de Providencia. **Caché 1 min.** Incluye totales de personas, empresas y estados de contactos.",
                "operationId": "getEstadisticas",
                "responses": {
                    "200": {
                        "description": "Estadísticas generadas (caché 1 min)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Estadisticas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ContactoInput": {
                "title": "ContactoInput",
                "required": [
                    "empresa_id",
                    "persona_id"
                ],
                "properties": {
                    "empresa_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "persona_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "notas_admin": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ContactoSolicitado": {
                "title": "ContactoSolicitado",
                "description": "Solicitud de contacto entre empresa y talento",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "empresa_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "persona_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "pendiente",
                            "contactado",
                            "entrevista",
                            "seleccionado",
                            "no-seleccionado",
                            "proceso-cerrado"
                        ]
                    },
                    "notas_admin": {
                        "type": "string",
                        "nullable": true
                    },
                    "fecha_contacto": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "fecha_entrevista": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "fecha_resultado": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ErrorGeneral": {
                "title": "Error General",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Recurso no encontrado."
                    }
                },
                "type": "object"
            },
            "ErrorValidacion": {
                "title": "Error de Validación",
                "description": "Error con detalle de campos inválidos (HTTP 422)",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Los datos enviados no son válidos."
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Estadisticas": {
                "title": "Estadísticas",
                "properties": {
                    "total_personas": {
                        "type": "integer",
                        "example": 45
                    },
                    "personas_validadas": {
                        "type": "integer",
                        "example": 38
                    },
                    "total_empresas": {
                        "type": "integer",
                        "example": 12
                    },
                    "empresas_validadas": {
                        "type": "integer",
                        "example": 10
                    },
                    "contactos_pendientes": {
                        "type": "integer",
                        "example": 5
                    },
                    "contactos_en_proceso": {
                        "type": "integer",
                        "example": 8
                    },
                    "contactos_exitosos": {
                        "type": "integer",
                        "example": 15
                    }
                },
                "type": "object"
            },
            "EstadoInput": {
                "title": "EstadoInput",
                "required": [
                    "estado"
                ],
                "properties": {
                    "estado": {
                        "type": "string",
                        "enum": [
                            "pendiente",
                            "contactado",
                            "entrevista",
                            "seleccionado",
                            "no-seleccionado",
                            "proceso-cerrado"
                        ]
                    },
                    "notas_admin": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "NotFound": {
                "description": "Recurso no encontrado",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorGeneral"
                        }
                    }
                }
            },
            "ValidationError": {
                "description": "Error de validación (HTTP 422)",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorValidacion"
                        }
                    }
                }
            }
        }
    },
...
}